requirements.txt
sounddevice==0.3.10
soundfile==0.10.2
ffmpeg-python==0.1.11
PyQt5==5.10.1
numpy==1.14.3
pyyaml==3.12